+++ src/components/Share/components/posters/holder.tsx
@@ -2,7 +2,7 @@ import { Box, Flex, Image } from '@chakra-ui/react'
 import { useRef } from 'react'
 import UserBgPath from '../../../../assets/share/bg/user.png'
 import { usePosterLoader } from '../../hooks/usePosterLoader'
-import { PosterProps } from '../../poster.interface'
+import { PosterProps } from '../../share.interface'
 import { Footer } from '../footer'
 import { useUrlToBase64 } from '../../../../hooks/useUrlToBase64'
 import { useTranslation } from 'react-i18next'
