--- src/components/Share/poster.interface.ts
@@ -1,4 +1,0 @@
-export interface PosterProps {
-  onLoaded: (el: HTMLDivElement) => void
-  shareUrl?: string
-}
