+++ src/views/Issuer/components/appbar.tsx
@@ -6,18 +6,26 @@ import {
   Appbar as RowAppbar,
   AppbarButton,
 } from '../../../components/Appbar'
-import { PosterType, Share } from '../../../components/Share/next'
+import { PosterType } from '../../../components/Share/share.interface'
 import { useHistoryBack } from '../../../hooks/useHistoryBack'
 import { formatCount } from '../../../utils'
 import { ReactComponent as BackSvg } from '../../../assets/svg/back.svg'
 import { ReactComponent as ShareSvg } from '../../../assets/svg/share.svg'
 import { useIssuerInfo } from '../hooks/useIssuerInfo'
 import { useShareImage } from '../hooks/useShareImage'
+import { lazy } from 'react'
+import { LoadableComponent } from '../../../components/GlobalLoader'
+
+const Share = lazy(async () => await import('../../../components/Share'))
 
 export const Appbar: React.FC = () => {
   const goBack = useHistoryBack()
   const { i18n } = useTranslation('translations')
-  const { isOpen, onOpen, onClose } = useDisclosure()
+  const {
+    isOpen: isOpenShare,
+    onOpen: onOpenShare,
+    onClose: onCloseShare,
+  } = useDisclosure()
   const { id } = useParams<{ id: string }>()
   const { data } = useIssuerInfo(id)
   const [shareImage] = useShareImage()
@@ -32,32 +40,34 @@ export const Appbar: React.FC = () => {
             </AppbarButton>
           }
           right={
-            <AppbarButton transparent onClick={onOpen}>
+            <AppbarButton transparent onClick={onOpenShare}>
               <ShareSvg />
             </AppbarButton>
           }
         />
       </AppbarSticky>
-      {data && (
-        <Share
-          isOpen={isOpen}
-          onClose={onClose}
-          shareUrl={window.location.href}
-          poster={{
-            type: PosterType.Issuer,
-            data: {
-              username: data.name,
-              avatarUrl: data.avatar_url ?? '',
-              isVerified: data.verified_info?.is_verified,
-              verifiedTitle: data.verified_info?.verified_title,
-              desc: data.description ?? '',
-              follow: formatCount(data.issuer_follows, i18n.language),
-              like: formatCount(data.issuer_likes, i18n.language),
-              coverImage: shareImage,
-            },
-          }}
-        />
-      )}
+      {data ? (
+        <LoadableComponent>
+          <Share
+            isOpen={isOpenShare}
+            onClose={onCloseShare}
+            shareUrl={window.location.href}
+            poster={{
+              type: PosterType.Issuer,
+              data: {
+                username: data.name,
+                avatarUrl: data.avatar_url ?? '',
+                isVerified: data.verified_info?.is_verified,
+                verifiedTitle: data.verified_info?.verified_title,
+                desc: data.description ?? '',
+                follow: formatCount(data.issuer_follows, i18n.language),
+                like: formatCount(data.issuer_likes, i18n.language),
+                coverImage: shareImage,
+              },
+            }}
+          />
+        </LoadableComponent>
+      ) : null}
     </>
   )
 }
